Added MTLTexture and MTLStoreAction in types.d. Also changed most enum data-types to size_t. Added contents selector to MTLBuffer. Added makeDefaultLibrary to MTLDevice. Added texture to MTLDrawable. Fixed data types to size_t for MTLRenderCommandEncoder. Added texture and storeAction to MTLRenderPassColorAttachmentDescriptor. Added label to MTLRenderPipelineDescriptor.

+++ source/metalrendering/rendercommandencoder.d
@@ -11,7 +11,7 @@ extern(Objective-C)
 extern class MTLRenderCommandEncoder : NSObject
 {
     void setRenderPipelineState(MTLRenderPipelineState renderPipelineState) @selector("setRenderPipelineState:");
-    void setVertexBuffer(MTLBuffer vertexBuffer, uint offset, uint index) @selector("setVertexBuffer:offset:atIndex:");
-    void drawPrimitives(MTLPrimitiveType primitiveType, uint vertexStart, uint vertexCount) @selector("drawPrimitives:vertexStart:vertexCount:");
+    void setVertexBuffer(MTLBuffer vertexBuffer, size_t offset, size_t index) @selector("setVertexBuffer:offset:atIndex:");
+    void drawPrimitives(MTLPrimitiveType primitiveType, size_t vertexStart, size_t vertexCount) @selector("drawPrimitives:vertexStart:vertexCount:");
     void endEncoding() @selector("endEncoding");
 }
